Moving resolver
